Add support for oslo messaging notifications in Keystone charm
This change introduces a new configuration option `enable-oslo-notify`
to the Keystone charm, which allows users to enable `oslo_messaging_notifications`.
This is necessary for services like Fleio that require Keystone to send notifications.

If `enable-oslo-notify` is set to `True`, the following section is added
to the `keystone.conf`:
[oslo_messaging_notifications] driver = messagingv2

By default, `enable-oslo-notify` is set to `False`, keeping the existing behavior
unchanged unless explicitly configured.

Closes-Bug: #2085335
Change-Id: I0b294f0feea3cac25492e6a40fa662c81819f61e